requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
et-xmlfile==1.1.0
frozenlist==1.3.3
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
multidict==6.0.4
numpy==1.24.2
openai==0.27.2
openpyxl==3.1.2
pandas==1.5.3
python-dateutil==2.8.2
python-telegram-bot==20.2
pytz==2023.2
requests==2.28.2
rfc3986==1.5.0
scipy==1.10.1
six==1.16.0
sniffio==1.3.0
tenacity==8.2.2
tiktoken==0.4.0
tqdm==4.65.0
urllib3==1.26.15
yarl==1.8.2
